move program model to core app with no database changes
