+++ registrar/apps/core/models.py
@@ -7,6 +7,7 @@ from django.utils.translation import ugettext_lazy as _
 from guardian.shortcuts import remove_perm
 from model_utils.models import TimeStampedModel
 
+from registrar.apps.common.data import DiscoveryProgram
 from registrar.apps.core import permissions as perms
 
 
@@ -66,6 +67,49 @@ class Organization(TimeStampedModel):
 
     def __str__(self):
         return self.name  # pragma: no cover
+
+
+class Program(TimeStampedModel):
+    """
+    Table that referrences a course-discovery Program entity.
+
+    .. no_pii::
+    """
+    class Meta(object):
+        app_label = 'core'
+        db_table = 'enrollments_program'
+
+    key = models.CharField(unique=True, max_length=255)
+    discovery_uuid = models.UUIDField(db_index=True, null=True)
+    managing_organization = models.ForeignKey(Organization, related_name='programs')
+
+    @property
+    def discovery_program(self):
+        return DiscoveryProgram.get(self.discovery_uuid)
+
+    @property
+    def title(self):
+        return self._get_cached_field('title')
+
+    @property
+    def url(self):
+        return self._get_cached_field('url')
+
+    @property
+    def program_type(self):
+        return self._get_cached_field('program_type')
+
+    def _get_cached_field(self, field):
+        """
+        Returns the specified field from a cached Discovery program.
+        If the program is not found in the cache it is loaded from discovery
+        """
+        discovery_program = DiscoveryProgram.get(self.discovery_uuid)
+        val = getattr(discovery_program, field)
+        return val
+
+    def __str__(self):
+        return self.key  # pragma: no cover
 
 
 class OrganizationGroup(Group):
